python-worker/requirements.txt
fastapi
pydantic
pydub
elevenlabs
whisperx
torchaudio
bs4
trafilatura
openai
moviepy
uvicorn
python-dotenv
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
lxml_html_clean
numpy<2.0
